refactor: address review - default wallet to signer, hide deploy/attach
