+++ tests/unit/test_relayer_deploy_deposit_wallet.py
@@ -3,15 +3,13 @@ import asyncio
 from urllib.parse import urlparse
 
 import httpx
-import pytest
 from _relayer_helpers import (
     install_relayer_routes,
     make_deposit_client,
-    make_proxy_client,
     request_json,
 )
 
-from polymarket.errors import UserInputError
+from polymarket._internal.actions.relayer.gasless import submit_deposit_wallet_create
 
 
 def test_deploy_deposit_wallet_wire_payload_has_no_signature() -> None:
@@ -31,7 +29,7 @@ def test_deploy_deposit_wallet_wire_payload_has_no_signature() -> None:
             },
         )
         try:
-            handle = await client.deploy_deposit_wallet()
+            handle = await submit_deposit_wallet_create(client._ctx)
             assert handle.transaction_id == "tx-dep-create"
         finally:
             await client.close()
@@ -42,16 +40,4 @@ def test_deploy_deposit_wallet_wire_payload_has_no_signature() -> None:
     assert body["type"] == "WALLET-CREATE"
     assert "signature" not in body
     assert "nonce" not in body
-    assert body["metadata"] == "Deploy Deposit Wallet"
-
-
-def test_deploy_deposit_wallet_rejects_non_deposit_wallet() -> None:
-    async def run() -> None:
-        client = await make_proxy_client()
-        try:
-            with pytest.raises(UserInputError, match="DEPOSIT_WALLET"):
-                await client.deploy_deposit_wallet()
-        finally:
-            await client.close()
-
-    asyncio.run(run())
+    assert body["metadata"] == ""
